+++ example/counter/lib/main.dart
@@ -5,6 +5,7 @@ void main() {
   runApp(MyApp());
 }
 
+
 /// ViewModel of Counter
 class CounterState {
   CounterState({