+++ lib/src/core/extension/disposable.dart
@@ -1,7 +1,0 @@
-import 'package:stated/src/core/core.dart';
-
-extension DisposableDisposerExtension on Disposable {
-  void disposeBy(Disposer disposer) {
-    disposer.addDispose(dispose);
-  }
-}