+++ lib/src/core/extension/listenable.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/foundation.dart';
-
-extension ListenableExtension on Listenable {
-  /// returns closure to remove the listener ([callback])
-  VoidCallback subscribe(VoidCallback callback) {
-    addListener(callback);
-    return () => removeListener(callback);
-  }
-}