+++ lib/stated.dart
@@ -1,7 +1,6 @@
 library stated;
 
 export 'src/core/core.dart';
-export 'src/rx/rx.dart';
 export 'src/stated/future_stated_builder.dart';
 export 'src/stated/stated.dart';
 export 'src/stated/stated_builder.dart';